requirements.txt
torch>=2.0
torchvision>=0.15
PyQt5>=5.15
Pillow>=9.0
matplotlib>=3.5
numpy>=1.23
geomloss>=0.2
